refactor : Rolling 추상클래스로 변수 선언
